Test class created with mocking, Calculations for percentages changed from long to double, style for display adapted to changes, delete unused dependency 'JUnit', implemented dependency spring-boot-starter-test

+++ src/main/java/de/cominto/praktikum/Math4Juerina_Web/service/MathServices.java
@@ -24,14 +24,14 @@ public interface MathServices {
 	Task saveTask(Task task);
 	Collection<Task>findByLastRoundAndDay(long userId, long roundId, Date date);
 
-	List<Long>findAllTasksFromLastFiveRoundsInfrintAcutalRound(Round round, long playerId);
+	List<Double>findAllTasksFromLastFiveRoundsInfrintAcutalRound(Round round, long playerId);
 	
 	
 	double getCorrectPercent(long rundId);
 	
 	int getNumberOfErrors(long roundId);
 
-	long getPercentCorrectFromDateToLocalDate(long playerId, int priviousDays);
-	List<Long> getCountAllTaskFromDateToDateGroupByDay(long playerId, int priviousDays);
+	Double getPercentCorrectFromDateToLocalDate(long playerId, int priviousDays);
+	List<Double> getCountAllTaskFromDateToDateGroupByDay(long playerId, int priviousDays);
 
 }
